Handle all dynamic behaviour in manager instead of renderer
